+++ socketsecurity/__init__.py
@@ -1,3 +1,3 @@
 __author__ = 'socket.dev'
-__version__ = '2.6.11'
+__version__ = '2.6.12'
 USER_AGENT = f'SocketPythonCLI/{__version__}'